changes made to how images looks
Images will be now appear inside div tag as a background image instead of in a img tag. Also added a head box that shows in which category are you in and last but not least, fixed styling which wasn't appearing in category desktop page because I forgot to link it correctly.